style(OutputManager): 출력 로직을 하나의 매니저로 관리하게 변경
