+++ src/main/java/racingcar/manager/RaceManager.java
@@ -8,6 +8,7 @@ import racingcar.service.CarService;
 public class RaceManager {
 
     private InputManager inputManager;
+    private OutputManager outputManager;
     private CarFactory carFactory;
     private CarService carService;
     private List<Car> cars;
@@ -24,11 +25,12 @@ public class RaceManager {
     public void startRace() {
 
         cars = carFactory.generateCars();
+        outputManager = new OutputManager(cars);
         tryCount = Integer.parseInt(inputManager.inputTryCount());
 
         racing();
 
-        printWinner();
+        outputManager.printWinner();
     }
 
     private void racing() {
@@ -37,22 +39,9 @@ public class RaceManager {
             for (Car car : cars) {
                 carService.ForwardOrStop(car);
             }
-            printInterResult();
+            outputManager.printInterResult();
         }
     }
 
-    private void printInterResult() {
-        System.out.println("\n실행 결과");
-        for (Car car : cars) {
-            System.out.println(car.getState());
-        }
-    }
 
-    public void printWinner() {
-        Referee referee = new Referee(cars);
-        List<String> winners = referee.selectWinner();
-        String winnerMessage = String.join(", ", winners);
-
-        System.out.println("\n최종 우승자 : " + winnerMessage);
-    }
 }
